fix(ui): revert node type change of `metadata_accumulator` -> `core_metadata`
The type of this node was changed due to breaking changes in f0db4d36e459a802abe9843ce619cc9e9bd3e3e0.

The change to the node type means pending queue items using this node would break if the app was updated.

In the end, there were no breaking changes, so to prevent queue items being stale and erroring, we can safely revert the change.

Note: This node was hidden from the workflow editor, so nobody has used it there in any official release.